chore: cleanup of update transactions fx
